+++ packages/html-markdown/__test__/table.spec.ts
@@ -37,7 +37,11 @@ describe('renderTable', () => {
     ['RIGHT', 'right'],
     ['center', 'center'],
     [null, null],
+    [undefined, null],
     ['invalid', null],
+    ['__proto__', null],
+    ['constructor', null],
+    ['left" onclick="alert(1)', null],
   ])('renders header and body alignment %j', (align, expected) => {
     const html = render({
       type: 'table',
@@ -53,6 +57,45 @@ describe('renderTable', () => {
         `<th class="yozora-table-cell"${attribute}><span class="yozora-text">Heading</span></th>` +
         '</tr></thead><tbody class="yozora-table__tbody"><tr class="yozora-table-row">' +
         `<td class="yozora-table-cell"${attribute}><span class="yozora-text">Cell</span></td>` +
+        '</tr></tbody></table>',
+    )
+  })
+
+  it.each([
+    ['yozora_test_alignment', 'yozora_test_alignment', ''],
+    ['1', 'left', ' align="left"'],
+  ])('ignores polluted prototype property %j in alignment output', (property, align, expected) => {
+    const node = {
+      type: 'table',
+      columns: [{ align }],
+      children: ['Header', 'Body'].map(value => ({
+        type: 'tableRow',
+        children: [
+          { type: 'tableCell', children: [text(value)] },
+          { type: 'tableCell', children: [text('Extra')] },
+        ],
+      })),
+    }
+    const previous = Object.getOwnPropertyDescriptor(Object.prototype, property)
+    let html: string
+    Object.defineProperty(Object.prototype, property, {
+      configurable: true,
+      writable: true,
+      value: 'left" onclick="alert(1)',
+    })
+    try {
+      html = render(node)
+    } finally {
+      if (previous) Object.defineProperty(Object.prototype, property, previous)
+      else Reflect.deleteProperty(Object.prototype, property)
+    }
+    expect(html).toBe(
+      '<table class="yozora-table"><thead class="yozora-table__thead"><tr class="yozora-table-row">' +
+        `<th class="yozora-table-cell"${expected}><span class="yozora-text">Header</span></th>` +
+        '<th class="yozora-table-cell"><span class="yozora-text">Extra</span></th>' +
+        '</tr></thead><tbody class="yozora-table__tbody"><tr class="yozora-table-row">' +
+        `<td class="yozora-table-cell"${expected}><span class="yozora-text">Body</span></td>` +
+        '<td class="yozora-table-cell"><span class="yozora-text">Extra</span></td>' +
         '</tr></tbody></table>',
     )
   })