+++ packages/html-markdown/src/renderer/table.ts
@@ -1,11 +1,5 @@
-import type { AlignType, Table, TableCell, TableRow } from '@yozora/ast'
+import type { Table, TableCell, TableRow } from '@yozora/ast'
 import type { INodeRenderer } from '../types'
-
-const ColumnAlignTypeMap: Record<string, AlignType> = {
-  left: 'left',
-  right: 'right',
-  center: 'center',
-}
 
 /**
  * Render Yozora Markdown AST node `Table` into HTML string.
@@ -14,8 +8,8 @@ const ColumnAlignTypeMap: Record<string, AlignType> = {
  */
 export const renderTable: INodeRenderer<Table> = (node, context) => {
   const aligns = node.columns.map(col => {
-    if (col.align == null) return null
-    return ColumnAlignTypeMap[col.align.toLowerCase()] ?? null
+    const align = col.align?.toLowerCase()
+    return align === 'left' || align === 'right' || align === 'center' ? ` align="${align}"` : ''
   })
 
   const [tHeadRow, ...tBodyRows] = node.children
@@ -23,7 +17,7 @@ export const renderTable: INodeRenderer<Table> = (node, context) => {
     '<thead class="yozora-table__thead"><tr class="yozora-table-row">' +
     tHeadRow.children
       .map((th: TableCell, i): string => {
-        const align = aligns[i] == null ? '' : ` align="${aligns[i]}"`
+        const align = Object.hasOwn(aligns, i) ? aligns[i] : ''
         const ths: string = context.renderChildren(th.children)
         return `<th class="yozora-table-cell"${align}>${ths}</th>`
       })
@@ -36,7 +30,7 @@ export const renderTable: INodeRenderer<Table> = (node, context) => {
       .map((row: TableRow) => {
         const tds: string = row.children
           .map((td: TableCell, i) => {
-            const align = aligns[i] == null ? '' : ` align="${aligns[i]}"`
+            const align = Object.hasOwn(aligns, i) ? aligns[i] : ''
             const ths: string = context.renderChildren(td.children)
             return `<td class="yozora-table-cell"${align}>${ths}</td>`
           })